refactor : service와 model 분리 service class와 model class 분리 Resolves : #1
