+++ src/main/java/baseball/service/GameService.java
@@ -1,30 +1,16 @@
 package baseball.service;
 
+import baseball.model.Game;
 import baseball.utils.Constant;
-import camp.nextstep.edu.missionutils.Randoms;
-import java.util.ArrayList;
-import java.util.List;
 
 public class GameService {
 
     private InputValidator validator;
-    private List<Integer> randomNumbers;
+    private Game game;
 
     public GameService() {
         this.validator = new InputValidator();
-        this.randomNumbers = generateNumbers();
-    }
-
-    private List<Integer> generateNumbers() {
-        List<Integer> numbers = new ArrayList<>();
-        while (numbers.size() < Constant.RANDOM_NUMBER_SIZE) {
-            int randomNumber = Randoms.pickNumberInRange(Constant.RANDOM_NUMBER_START,
-                    Constant.RANDOM_NUMBER_END);
-            if (!numbers.contains(randomNumber)) {
-                numbers.add(randomNumber);
-            }
-        }
-        return numbers;
+        this.game = new Game();
     }
 
     public String validateInput(String userInput) {
@@ -33,27 +19,15 @@ public class GameService {
     }
 
     private String generateResultMessage(String userInput) {
-        int ballCount = Constant.COUNT_ZERO;
-        int strikeCount = Constant.COUNT_ZERO;
+        int ballCount = 0;
+        int strikeCount = 0;
 
-        for (int i = Constant.COUNT_ZERO; i < userInput.length(); i++) {
-            if (isStrike(userInput, i)) {
-                strikeCount++;
-            } else if (isBall(userInput, i)) {
-                ballCount++;
-            }
+        for (int i = 0; i < userInput.length(); i++) {
+            ballCount += game.calculateBall(userInput, i);
+            strikeCount += game.calculateStrike(userInput, i);
         }
 
         return getResultMessage(ballCount, strikeCount);
-    }
-
-    private boolean isStrike(String userInput, int index) {
-        return Character.getNumericValue(userInput.charAt(index)) == randomNumbers.get(index);
-    }
-
-    private boolean isBall(String userInput, int index) {
-        int number = Character.getNumericValue(userInput.charAt(index));
-        return randomNumbers.contains(number) && !isStrike(userInput, index);
     }
 
     private String getResultMessage(int ballCount, int strikeCount) {
